Use self-signed SSL cert for local development
Not sure how useful this is in practical terms, but it seems like a
good idea to use SSL if we can.

Also add a script to start rails. The command with the cert and key
params was getting too long and messy.

Part of the motivation to do this is because I'm working on a
Dockerfile for building a deployable image, and I want that to have
SSL from the start.

Note: The cert warnings are annoying. Should figure out how to skip
the SSL warning for *.tiddlyspot.local.

Setting this doesn't seem to work:
  chrome://flags/#allow-insecure-localhost

This might work but I didn't try it yet:
  https://peacocksoftware.com/blog/make-chrome-auto-accept-your-self-signed-certificate