Modify certain tiddlers when serving sites
The important one is $:/UploadURL, but also set the other two that
we previously set on site creation.

This should mean you can rename your site and it will still be
correctly configured for saving, fixing issue #19.

Also:
- Refactor TiddlyWiki file handling and write some
  nicer code for parsing and injecting tiddlers with
  Nokogiri and some test coverage.
- Since we customize the site html when it's rendered
  we no longer need to customize it on site creation,
  so remove that.

It feels a little risky to be modifying TiddlyWiki files like this,
so we'll proceed cautiously.