Handle options request gracefully
Previously we hit a routing error that generated a 404 reponse. Now
we'll explicitly route it to a 404 response with no body.

This doesn't change TW's behaviour since it only looks at the
status, but it's better if we handle it nicely.

It appears rails checks the authenticity token for options requests,
so we needed to skip it to avoid an
ActionController::InvalidAuthenticityToken exception.

Issue #52