Introduce new version of UploadPlugin for Classic
Inject the plugin automatically when serving a TW Classic site.
There's a (good old) ERB based templating system for plugins so we
can conveniently include site specific things like the site name and
upload url.

The plugin is based on the original Bidix plugin of course, but I
removed some of the parameters we don't use and some features we
don't need.

Stuff removed includes the backstage extension, support for writing
RSS files, creation backups, plus the additional parameters to
upload plugin that we generally ignore.

Using a different namespace so we don't clash with the real
UploadPlugin if it's present.

I've resisted the temptation to add any UI tweaks and features. That
could come later. At this stage the goal is the simplest thing that
works.