Extract 'default' methods into a concern
Use it for plans and empties.

Also a tangentially related improvement to how the default empty is
accessed in the sites controller.

Note the default_id method isn't used anywhere yet, but I decided to
create it anyway.

Since there's only class methods, I briefly considered making it a
plain module and using extend, but nah - let's stick with the
delicious Rails koolaid. ;)