Allow the admin user to view private sites
It seems reasonable, but up until now the admin user wasn't
permitted to view a private site directly. Since the admin user
presumably can access all the data in S3, I guess this is not a big
deal, but it does makes it more convenient. To make it easier to
notice the site is private, make the link smaller in the admin site
list.

The motivation is that a user reported their site was broken and
asked me to look at it, but I couldn't because it was private.