Support filtering sites in admin pages
Also tweak which columns are shown depending on the filter options,
and use a shorter param for user filters just to make long query
strings slightly prettier.

So far all the filtering is based on "include/exclude/disregard"
using a boolean expression. Need to extend it to value based
filtering and name searches, but that can come later.