Rails runner script to remove dupe attachments
I noticed the dupes a while ago, but thought they were harmless. Now
I have reason to believe otherwise. It appears rails sometimes use one
for saving and another for serving, so if you reload the site it looks
like your save didn't work.

Also show a count of them in the admin page in case they
reappear.

I used a version of this to clean up 8 or 10 duplicates in
production.

Issue #88