Add endpoint for tiddlers in .tid format
Notice there's some caching. It's not as significant as the site
download caching in the previous commit, since the expensive
download is already cached, but still it seems like a good idea,
particularly if we imagine some popular tiddler content being
fetched many times.

See also the discussion in:
Issue #116