Some refactor and clean up in TwFile
Originally I had TiddlerDiv inheriting from Nokogiri::XML::Node, but
it turned out to be not that useful.

I also spent some time extending this to be compatible with the
older tiddler format, but that's on hold now that I realised
thost_upload_plugin would also need rewriting for compatibility.