Fix bug where searching in hub would throw errors
Fixes a regression introduced indirectly by the last patch, (which
was live in prod for a few hours).

It was an SQL error caused because 'name' is ambigous in `name ILIKE
..`. Join to blobs in the last patch caused the bug since the
attachments table also has a name field, but the root cause is
related to not using the appliable model search scopes correctly in
hub_query, since they do explicitly include the table name.

(This highlights the need for some better test coverage for the hub
controller I guess.)