Methods for bulk migrating and compressing
1. ensure_migrated
Takes an existing tspot and re-saves its content to an active
storage blob. Previously this was only done when a tspot site was
saved.

2. ensure_compressed
Re-saves an uncompressed site in order to compress it, but preserves
the blob created timestamp.

In hindsight these could have been defined in a fixup script rather
than in the model. Maybe later the methods can be deleted, but it
seems okay to leave them there for now.

Issue #135