Stop checking for new tspot sites
Also stop creating records for tspot sites that don't exist.

The reason for doing this is so the 404s for non-existing tiddylspot
sites are cheap and fast, and no longer require doing a lookup in
the dreamhost bucket. And that means I can ditch the records for
non-existing tiddylspot sites (see upcoming commits) and delete a
bunch of code.

Required some hacking of the test coverage which is starting to feel
a bit stretched due to the number of changes the tspot site handling
has been through. Fix it up later maybe.

Also add a missing assert_mock in claim_sites_test.