Support fetching tiddler data from new json store
Closely related to the previous commit, but this is about reading
specific tiddlers from the new format script element with json
inside it. This is needed so the /text/Tiddler.tid and the
/tiddlers.json end-points work for the new format store.

Do some extra work to support multiple store elements for
compatibility with my understanding of TW's own behavior, and also
because we are using them ourselves to inject tiddlers, (see
previous commit).

I have some concerns about the memory requirements, since it's
likely we'll be loading the entire store including the very large
"$:/core" tiddler into memory multiple times, but I guess we'll see
how it goes.

Issue #161