Refactor the store accessors in TwFile
Mainly to ensure the tiddler data is refreshed after you add to it.

It might not matter in production, but it's better if this works
in a sane way. When you add content to a TiddlyWiki you can read it
back. Also it fixes some failing tests if you use a 5.2+ empty file
when running the test suite.

Could have been done without the refactor, but it seems better this
way. Use some memoized accessors so we can just clear the relevant
instance vars when we need to refresh.

Note: (@_things || doc.xpath(...)).first looks odd but it's done
like that so we don't look it up every time in the case where the
element doesn't exist. Because doc.xpath will return [] which is
truthy in ruby, the lazy memoization pattern will work like it
should.

Issue #161