Add stripe gem and pay gem plus related schema
The migrations are generated automatically by the pay gem - see the
pay gem installation docs for details.

Later: Notice that the timestamp on the migration and on this
commit's creation date is really old. It's because I've had this in
a wip branch for quite a while.

Later again: Decided to merge this change now even though the gems
and the new schema won't be used right away. Why? I'd like to not
have to continue rebasing the large migration and juggling the
schema differences between the main branch and the wip branch. Why
now and not sooner? The subscription functionality is mostly done
now, afaict the pay gem works pretty well, and the wip branch is
getting pretty stable.