Insert tiddlers related to being logged in
The motivation is so you can show a public site in "read only" mode
when the site is viewed by someone other than its owner.

Note we only include it when the logged in user is the site owner,
since that fits with the main use case.

Clearing UploadUrl when the user is not logged in is an interesting
idea, but I won't do it now, partly because it breaks get_site_name
which causes some other tests to fail.

Issue #189