Add an option to suppress X-Frame-Options header
Checking this option allows the site to be allowed to be used inside
an iframe from elsewhere on the internet. There are some small but
non-zero clickjacking risks involved, so that's why you have to
explicitly enable it.

For tiddlyspot sites, the lack of SSL is probably also going to
cause browsers to block loading in an iframes, but I'm not going to
spend time testing that.

Also introduces a concern for the similar controllers and a shared
partial directory for common form elements. (I'll add more partials
there later probably.)

I didn't add test coverage for the tspot site but presumably if it
works for one it should work for the other.

Closes #206