Add some buttons next to actions menu
Maybe I'm not done tweaking the actions menu and buttons, but this
seems like an improvement.

Button actions are duplicated in the menu intentionally since I
think the icons alone are not enough to communicate what the action
is, especially on touch screen where there is no mouseover to show
the title.