Introduce site option to enable put saver
I chose not to support this for Tiddlyspot sites. There's no
technical reason not to I guess, but I don't want to invest time
adding and testing this kind of new feature for legacy Tiddlyspot
sites.

Rolling it out as an option allows it to be soft-launched as a kind
of tech preview. Maybe in future it will become the default saving
method, but not immediately.

The option doesn't do anything yet, but see next commit.