>>> Repo split! <<<
Every commit older than this one was originally made in the
tiddlyhost repo[1], and all the files in the top level here were
under the 'rails' directory in the original repo.

The history from the older repo was rewritten using a filter-branch
command[2]. It was easier than I thought it would be, and despite
the dire warnings suggesting to use filter-repo instead, it seems to
have worked pretty well.

[1] git@github.com:simonbaird/tiddlyhost.git
[2] git filter-branch --subdirectory-filter rails