Add support for hosting FeatherWiki sites
FeatherWiki is super new, but it seems fun and full of potential, so
I wanted to try adding a TiddlyWiki style put saver and making it
work on Tiddlyhost.

To begin with we're using "Dove" 1.2.1 modified to include the put
saver. Later we'll use "Tern" which has the put saver already.

See also:
* https://feather.wiki/
* https://codeberg.org/Alamantus/FeatherWiki/pulls/5