Nicer email confirmation and password reset emails
Prettify and rework the emails sent by devise, including signup
confirmation, password reset, and a few others. Include some
branding and improve the text.

Up until now we were using the standard default devise emails which
are very generic and not very good.

Introduce the bootstrap-mailer gem for generating email client
compatible html because that is very hard to do apparently, and
convert the erb to haml.

Also improve the after_registration page and include some
troubleshooting advice on it.

While testing this I learned that the "change email" confirmation
email is the same as the signup confirmation, so I added some logic
so the wording varies between the two.

Also I learned that some of these emails seem to be unused. Will try
to confirm that and maybe remove them in future.

Closes #3