Fix change email confirmation recipient bug
When I added the "pretty" email address it caused the confirmation
email when changing email address to be sent incorrectly to the old
address instead of the new one.

Rather than try to fix it, (probably by using the unconfirmed_email
field instead of the email field as and when required, which seems
like more effort than it's worth), let's just ditch the pretty email
altogether.