Get grover & puppeteer working in base image
Puppeteer uses a headless chromium which has some dependencies we
can be provided by installing chrome, so do that.

The default seccomp policy prevents chrome sandbox for working.
There are numerous suggested ways to fix that, but I was having
problems making it work consistently, so I'll start with
"unconfined" and try to figure it out a stricter policy in a future
commit.

See also:
* https://github.com/puppeteer/puppeteer/blob/main/docs/troubleshooting.md#setting-up-chrome-linux-sandbox
* https://github.com/jessfraz/dockerfiles/issues/65#issuecomment-215224035
* https://github.com/moby/moby/blob/master/profiles/seccomp/default.json
* https://docs.docker.com/engine/security/seccomp/
* https://kubernetes.io/docs/tutorials/security/seccomp/