Fix email link test fail introduced by gem update
My assumption is that there were some changes in the bootstrap-email
gem in the last bundle update that caused it to start creating a
text version of the emails automatically generated from the html
content. This caused a test failure which is fixed by this patch.

Also, change `..body.encoded` to `..body.decoded` which I think it
should have been all along.