Update classic empty to version 2.9.3
Looks like the first release in over three years.

Required a tweak in lib/tw_file to make the version detection regex
work for the new version. (I'm glad the test suite was good enough
to catch that.)

Keep the old empty checked in since it will be included in tests,
(see the `for_all_empties` method).

See also:
https://github.com/TiddlyWiki/TiddlyWiki/pull/274#issuecomment-1166461319